Remove condition as portion of id
Parsing for the item condition is currently unstable. The issue is with the
server side not the parser. We could grab for an additional detail html site
per item(!), which seems more accurate. Since this is rather expensive, lets
compare the known outlet items based on name and price only.